% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumnOnesDS.R
\name{addColumnOnesDS}
\alias{addColumnOnesDS}
\title{addColumnOnesDS}
\usage{
addColumnOnesDS(X, columns)
}
\arguments{
\item{X}{A string representing a dataframe}

\item{columns}{A character vector of column names for the new dataframe}
}
\value{
A dataframe with a column of ones added
}
\description{
The function adds a column of ones to a dataframe and returns the new dataframe.
}
