% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppendInfluenceDS.R
\name{AppendInfluenceDS}
\alias{AppendInfluenceDS}
\title{AppendInfluenceDS}
\usage{
AppendInfluenceDS(df, influences, id_period_vector, column)
}
\arguments{
\item{df}{A string representing a dataframe}

\item{influences}{A vector of influence values}

\item{id_period_vector}{A vector of row names or a string representing a vector of row names}

\item{column}{A character string representing the name of the column to append the influence values to}

\item{seed}{A seed used for random sampling}
}
\value{
A dataframe with the column of influence values added
}
\description{
The function appends a column of influence values to a dataframe and returns the new dataframe.
}
