% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHelperDS.R
\name{buildHelperDS}
\alias{buildHelperDS}
\title{assign function -- provide estimate of found coefficients (typically from federated RDataShield GLM function)}
\usage{
buildHelperDS(x)
}
\arguments{
\item{form}{formula in text format}

\item{coefficients}{coefficients of dependent variables in form}

\item{object}{dataframe object name where dependent var values are located}

\item{invlog}{use inverse of logit}
}
\value{
estimate/prediction using regression coefficients
}
\description{
assign function -- provide estimate of found coefficients (typically from federated RDataShield GLM function)
}
