% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterInfluenceFunctionDS.R
\name{clusterInfluenceFunctionDS}
\alias{clusterInfluenceFunctionDS}
\title{clusterInfluenceFunctionDS}
\usage{
clusterInfluenceFunctionDS(df, influence_matrix, clustervars, idname)
}
\arguments{
\item{df}{A string representing a dataframe}

\item{influence_matrix}{A matrix of influence values}

\item{clustervars}{A character string or vector of strings representing the column(s) in the dataframe that define the clusters}

\item{idname}{A character string representing the column in the dataframe that contains the unique identifier for each observation}
}
\value{
A vector of mean influence values for each cluster
}
\description{
The function calculates the mean influence for each cluster in a dataframe and returns the values.
}
