% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMatrixCrossproductDS.R
\name{computeMatrixCrossproductDS}
\alias{computeMatrixCrossproductDS}
\title{computeMatrixCrossproductDS}
\usage{
computeMatrixCrossproductDS(X)
}
\arguments{
\item{X}{A string representing a matrix}
}
\value{
A matrix representing the cross-product of the input matrix
}
\description{
The function computes the cross-product of a matrix and returns the result.
}
