% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEmptyIdMatrixDS.R
\name{createEmptyIdMatrixDS}
\alias{createEmptyIdMatrixDS}
\title{createEmptyIdMatrixDS}
\usage{
createEmptyIdMatrixDS(df, idname, n_columns)
}
\arguments{
\item{df}{A dataframe or a string representing a dataframe}

\item{idname}{A character string representing the column in the dataframe that contains the unique identifier for each observation}

\item{n_columns}{A numeric value or a string representing the number of columns in the empty matrix}
}
\value{
An empty matrix with unique id values as row names and a specified number of columns
}
\description{
The function creates an empty matrix with unique id values as row names and a specified number of columns.
}
