% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enoughIndividualsDS.R
\name{enoughIndividualsDS}
\alias{enoughIndividualsDS}
\title{assign function -- provide estimate of found coefficients (typically from federated RDataShield GLM function)}
\usage{
enoughIndividualsDS(df, colname, value)
}
\arguments{
\item{value}{}
}
\value{
estimate/prediction using regression coefficients
}
\description{
assign function -- provide estimate of found coefficients (typically from federated RDataShield GLM function)
}
