% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDfSubsetDS.R
\name{genDfSubsetDS}
\alias{genDfSubsetDS}
\title{genDfSubsetDS}
\usage{
genDfSubsetDS(df, name_variable, value, include_zero = TRUE)
}
\arguments{
\item{df}{A dataframe or a string representing a dataframe}

\item{name_variable}{A character string representing the column in the dataframe to subset on}

\item{value}{A value or a string representing the value to subset on in the specified column}

\item{include_zero}{A boolean value indicating whether or not to include rows with a value of zero in the subset}
}
\value{
A subset of the dataframe with rows that match the specified variable and value
}
\description{
The function generates a subset of a dataframe based on a specified variable and value.
}
