% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPropDS.R
\name{genPropDS}
\alias{genPropDS}
\title{assign function -- provide estimate of found coefficients (typically from federated RDataShield GLM function)}
\usage{
genPropDS(form, coefficents, object, invlog, constant_in_matrix = FALSE)
}
\arguments{
\item{form}{formula in text format}

\item{object}{dataframe object name where dependent var values are located}

\item{invlog}{use inverse of logit}

\item{coefficients}{coefficients of dependent variables in form}
}
\value{
estimate/prediction using regression coefficients
}
\description{
assign function -- provide estimate of found coefficients (typically from federated RDataShield GLM function)
}
