% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplierBootstrapDS.R
\name{multiplierBootstrapDS}
\alias{multiplierBootstrapDS}
\title{multiplierBootstrapDS}
\usage{
multiplierBootstrapDS(matrix, n_iterations)
}
\arguments{
\item{matrix}{A matrix or a string representing a matrix}

\item{n_iterations}{A numeric value or a string representing the number of iterations for the bootstrap}
}
\value{
A matrix that is the result of the multiplier bootstrap
}
\description{
The function performs the multiplier bootstrap on a given matrix and number of iterations, and returns the bootstrapped matrix.
}
