% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplyMatrixMatrixDS.R
\name{multiplyMatrixMatrixDS}
\alias{multiplyMatrixMatrixDS}
\title{multiplyMatrixMatrixDS}
\usage{
multiplyMatrixMatrixDS(
  matrix1,
  matrix2,
  nrow1 = NULL,
  ncol1 = NULL,
  nrow2 = NULL,
  ncol2 = NULL
)
}
\arguments{
\item{matrix1}{A matrix or a string representing a matrix}

\item{matrix2}{A matrix or a string representing a matrix}

\item{nrow1}{The number of rows for the first matrix, used only if the first matrix is not a string}

\item{ncol1}{The number of columns for the first matrix, used only if the first matrix is not a string}

\item{nrow2}{The number of rows for the second matrix, used only if the second matrix is not a string}

\item{ncol2}{The number of columns for the second matrix, used only if the second matrix is not a string}
}
\value{
A matrix representing the product of the two input matrices
}
\description{
The function multiplies two matrices and returns the product matrix. At least one of the matrices must be from the server-side.
}
