% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplyMatrixScalarDS.R
\name{multiplyMatrixScalarDS}
\alias{multiplyMatrixScalarDS}
\title{multiplyMatrixScalarDS}
\usage{
multiplyMatrixScalarDS(matrix, scalar)
}
\arguments{
\item{matrix}{The matrix to be multiplied.  Must be a string with the name of the matrix on the server side.}

\item{scalar}{The scalar to be multiplied. Must be a single number.}
}
\value{
A matrix representing the product of the two inputs
}
\description{
The function is used to multiply a matrix and a scalare.
}
