% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeDS.R
\name{recodeDS}
\alias{recodeDS}
\title{recodeDS}
\usage{
recodeDS(x, replace_with, replace)
}
\arguments{
\item{x}{Vector in which values should be replaced}

\item{replace_with}{new value}

\item{replace}{Entry to be replaced}
}
\value{
The recoded vector
}
\description{
The function creates a new object of the given vector and replaces certain entries of this vector
}
