% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeOddsDS.R
\name{computeOddsDS}
\alias{computeOddsDS}
\title{Computes odds from propensities and a dummy variable}
\usage{
computeOddsDS(name_propensities, name_dummy_g)
}
\arguments{
\item{name_propensities:}{A vector of propensities represented as a string}

\item{name_dummy_g:}{A vector of dummy variables representing a binary outcome represented as a string}
}
\value{
: A vector of odds computed from the propensities and the binary outcome
}
\description{
Computes odds from propensities and a dummy variable
}
