% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enoughIndividualsDS.R
\name{enoughIndividualsDS}
\alias{enoughIndividualsDS}
\title{enoughIndividualsDS}
\usage{
enoughIndividualsDS(df, colname, value)
}
\arguments{
\item{df}{A dataframe or a string representing a dataframe}

\item{colname}{A character string representing the column in the dataframe to check for the specified value}

\item{value}{A value or a string representing the value to check for in the specified column}
}
\value{
A boolean indicating whether or not there are enough individuals with the specified value in the specified column.
}
\description{
The function checks if there are enough individuals with a specified value in a specified column of a dataframe.
}
