% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNotYetTreatedDS.R
\name{generateNotYetTreatedDS}
\alias{generateNotYetTreatedDS}
\title{generateNotYetTreatedDS}
\usage{
generateNotYetTreatedDS(df, name_variable, t, g)
}
\arguments{
\item{df}{A dataframe or a string representing a dataframe}

\item{name_variable}{A character string representing the column in the dataframe to subset on}

\item{t}{An integer representing the minimum period value}

\item{g}{An integer representing the current period}
}
\value{
A subset of the dataframe with rows that match the specified conditions
}
\description{
The function generates a subset of a dataframe based on a specified variable and value, while only including observations that are from the g period or after, or have never been treated.
}
