% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPropDS.R
\name{genPropDS}
\alias{genPropDS}
\title{genPropDS}
\usage{
genPropDS(form, coefficents, object, invlog, constant_in_matrix = FALSE)
}
\arguments{
\item{form}{A formula in character string format.}

\item{object}{An R object that contains the variables in the formula}

\item{invlog}{A logical value indicating whether to apply inverse logit transformation on the estimate}

\item{constant_in_matrix}{A logical value indicating whether the constant term is included in the coefficient matrix}

\item{coefficients}{A numeric vector of coefficients for the variables in the formula}
}
\value{
A numeric value representing the estimate of the variable
}
\description{
The function calculates the estimate of a variable based on the given formula, coefficients and object, and returns the numeric value of the variable.
}
